feat: transition level design
